requirements.txt
Django
djangorestframework
Pyannite
